♻️ [desktop]: extract log plugin into logging module and use KeepSome(3)
- Moved log_plugin_builder and log_file_target into src/logging.rs
- Replaced RotationStrategy::KeepAll with KeepSome(3) so the plugin
  handles log rotation natively
- Removed manual cleanup_old_logs function (now redundant)
- Moved startup log::info! into setup hook where the logger is active
